third_party/blink/web_tests/virtual/layout_ng_svg_text/svg/custom/use-nested-missing-target-removed-expected.txt

CONSOLE ERROR: line 7: Uncaught TypeError: Cannot set properties of null (setting 'id')
PASS
